requirements.txt
pandas
numpy
requests
schedule
python-dotenv